fix(compiler-spec): extract instruction spec utilities
