+++ packages/compiler-spec/src/index.ts
@@ -8,6 +8,7 @@ export * from './errors';
 export * from './functionTypes';
 export * from './instructionSpecs';
 export * from './instructionSpecTypes';
+export * from './instructionSpecUtils';
 export * from './instructions';
 export * from './macros';
 export * from './memory';
